add visual crossing api client
Visual crossing is an api weather service, chosen primarily because it has a free tier (1000 daily requests) and it will provide a weather forecast based on a zip code alone.

The API key is stored in the encrypted file and can be accessed using Rails.application.credentials.visual_crossing_api_key.

Faraday was added as a versitle http client that supports many adapters andscales well if necessary.

factory_bot was removed because we will not be using ActiveRecord in this app.

webmock was added to stub out requests organically