add forecasts without caching
Added and styled index page and show page which fill in an extended forecast
for a particular zip code.

The user puts in their address. When they submit an html form that includes
their zip code, turbo loads a show page with the details of the forecast.

Currently the forecast includes high temp, low temp, current conditions, and
five days of extended forecast.