add valkey (redis) caching support
When a user submits their address and we use their zipcode to get a
weather forcast, we process the response to get the current temperature,
current weather conditions, and an extended forecast that includes the
high temp, low temp, and expected conditions.

That information is stored in redis as a json that is mapped to the
user's zipcode. Any other user that shares the same zipcode will
receive a cached forecast if it is requested during the cache
expiration time.

If the weather result is returned from the cache, a banner will
notify the user that the data may be stale.

Note that we do not create a second database for the test enviornment.
The data that is stored in redis is meant to be short lived, so
running the test suite will also clear the cache for the development
environment. This would not affect production.